refactor(sidebar): rename sidebar-option-input to sidebar-input-option; Add edit group option into group item dropdown
